removed pgcmosaic class, added GdalRaster
