+++ packages/geo/CMakeLists.txt
@@ -20,6 +20,7 @@ if (GDAL_FOUND AND PROJ_FOUND)
     target_sources(slideruleLib
         PRIVATE
             ${CMAKE_CURRENT_LIST_DIR}/geo.cpp
+            ${CMAKE_CURRENT_LIST_DIR}/GdalRaster.cpp
             ${CMAKE_CURRENT_LIST_DIR}/GeoRaster.cpp
             ${CMAKE_CURRENT_LIST_DIR}/GeoIndexedRaster.cpp
             ${CMAKE_CURRENT_LIST_DIR}/GeoJsonRaster.cpp
@@ -38,6 +39,7 @@ if (GDAL_FOUND AND PROJ_FOUND)
     install (
         FILES
             ${CMAKE_CURRENT_LIST_DIR}/geo.h
+            ${CMAKE_CURRENT_LIST_DIR}/GdalRaster.h
             ${CMAKE_CURRENT_LIST_DIR}/GeoRaster.h
             ${CMAKE_CURRENT_LIST_DIR}/GeoIndexedRaster.h
             ${CMAKE_CURRENT_LIST_DIR}/GeoJsonRaster.h
