+++ packages/geo/GeoRaster.h
@@ -36,40 +36,16 @@
  * INCLUDES
  ******************************************************************************/
 
+#include "GdalRaster.h"
 #include "RasterObject.h"
 #include "OsApi.h"
 #include "TimeLib.h"
-#include "GeoParms.h"
 #include "Ordering.h"
 #include "List.h"
-#include <ogr_geometry.h>
-#include <ogrsf_frmts.h>
 
 /******************************************************************************
  * Typedef and macros used by geo package
  ******************************************************************************/
-
-#define CHECKPTR(p)                                                           \
-do                                                                            \
-{                                                                             \
-    if ((p) == NULL)                                                          \
-    {                                                                         \
-        throw RunTimeException(CRITICAL, RTE_ERROR,                           \
-              "NULL pointer detected (%s():%d)", __FUNCTION__, __LINE__);     \
-    }                                                                         \
-} while (0)
-
-
-#define CHECK_GDALERR(e)                                                      \
-do                                                                            \
-{                                                                             \
-    if ((e))   /* CPLErr and OGRErr types have 0 for no error  */             \
-    {                                                                         \
-        throw RunTimeException(CRITICAL, RTE_ERROR,                           \
-              "GDAL ERROR detected: %d (%s():%d)", e, __FUNCTION__, __LINE__);\
-    }                                                                         \
-} while (0)
-
 
 
 
@@ -77,7 +53,7 @@ do                                                                            \
  * GEO RASTER CLASS
  ******************************************************************************/
 
-class GeoRaster: public RasterObject
+class GeoRaster: public RasterObject, public GdalRaster
 {
     public:
 
@@ -85,93 +61,12 @@ class GeoRaster: public RasterObject
          * Constants
          *--------------------------------------------------------------------*/
 
-        static const int INVALID_SAMPLE_VALUE          = -1000000;
-        static const int MAX_SAMPLING_RADIUS_IN_PIXELS = 50;
-        static const int SLIDERULE_EPSG                = 7912;
-
-        static const char* FLAGS_RASTER_TAG;
-        static const char* SAMPLES_RASTER_TAG;
-
-        /*--------------------------------------------------------------------
-         * Typedefs
-         *--------------------------------------------------------------------*/
-
-        typedef struct
-        {
-            double lon_min;
-            double lat_min;
-            double lon_max;
-            double lat_max;
-        } bbox_t;
-
-        class CoordTransform
-        {
-        public:
-            OGRCoordinateTransformation *transf;
-            OGRSpatialReference source;
-            OGRSpatialReference target;
-
-            void clear(bool close = true);
-            CoordTransform(void) { clear(false); }
-           ~CoordTransform(void) { clear(); }
-        };
-
-        class GeoDataSet
-        {
-        public:
-            CoordTransform  cord;
-            double          verticalShift;  /* Calculated for last POI transformed to target CRS */
-            std::string     fileName;
-            GDALDataset    *dset;
-            GDALRasterBand* band;
-            GDALDataType    dataType;
-            bool            dataIsElevation;
-            uint32_t        rows;
-            uint32_t        cols;
-            double          cellSize;
-            bbox_t          bbox;
-            int32_t         xBlockSize;
-            int32_t         yBlockSize;
-            uint32_t        radiusInPixels;
-
-            void clear(bool close = true);
-
-            GeoDataSet(void) { clear(false); }
-           ~GeoDataSet(void) { clear(); }
-        };
-
-        class Raster: public GeoDataSet
-        {
-        public:
-            bool            enabled;
-            bool            sampled;
-            std::string     groupId;
-
-            double          gpsTime;
-            double          useTime;
-
-            /* Last sample information */
-            OGRPoint        point;
-            sample_t        sample;
-
-            Raster(const char* _fileName="", double _gpsTime=0);
-
-            inline bool containsPoint(OGRPoint& p)
-            {
-                return (dset &&
-                        (p.getX() >= bbox.lon_min) && (p.getX() <= bbox.lon_max) &&
-                        (p.getY() >= bbox.lat_min) && (p.getY() <= bbox.lat_max));
-            }
-
-        };
-
-
         /*--------------------------------------------------------------------
          * Methods
          *--------------------------------------------------------------------*/
 
         virtual         ~GeoRaster  (void);
-        virtual void    getSamples  (double lon, double lat, double height, int64_t gps, List<sample_t>& slist, void* param=NULL) override;
+        virtual void    getSamples  (double lon, double lat, double height, int64_t gps, List<RasterObject::sample_t>& slist, void* param=NULL) override;
 
     protected:
 
@@ -179,19 +74,7 @@ class GeoRaster: public RasterObject
          * Methods
          *--------------------------------------------------------------------*/
 
-                        GeoRaster             (lua_State* L, GeoParms* _parms, const char* _fileName="", double _gpsTime=0);
-        void            openRaster            (const char* _fileName="", double _gpsTime=0);
-        void            openRaster            (Raster* raster);
-        virtual int64_t getRasterDate         (void);
-
-        const char*     getUUID               (char* uuid_str);
-        void            createTransform       (CoordTransform& cord, GDALDataset* dset);
-        virtual void    overrideTargetCRS     (OGRSpatialReference& target);
-        void            setCRSfromWkt         (OGRSpatialReference& sref, const std::string& wkt);
-        bool            containsWindow        (int col, int row, int maxCol, int maxRow, int windowSize);
-        int             radius2pixels         (double cellSize, int _radius);
-        void            readRasterWithRetry   (GDALRasterBand* band, int col, int row, int colSize, int rowSize,
-                                               void* data, int dataColSize, int dataRowSize, GDALRasterIOExtraArg *args);
+         GeoRaster  (lua_State* L, GeoParms* _parms, const char* _fileName="", double _gpsTime=0);
 
         /*--------------------------------------------------------------------
          * Data
@@ -208,7 +91,6 @@ class GeoRaster: public RasterObject
         /*--------------------------------------------------------------------
         * Data
         *--------------------------------------------------------------------*/
-        Raster _raster;
 
         /*--------------------------------------------------------------------
         * Methods
@@ -217,12 +99,6 @@ class GeoRaster: public RasterObject
         static int luaDimensions(lua_State* L);
         static int luaBoundingBox(lua_State* L);
         static int luaCellSize(lua_State* L);
-
-        void       readPOI           (Raster* raster);
-        void       readPixel         (Raster* raster);
-        void       resamplePixel     (Raster* raster);
-        void       computeZonalStats (Raster* raster);
-        bool       nodataCheck       (Raster* raster);
 };
