+++ plugins/pgc/plugin/ArcticDemMosaicRaster.h
@@ -36,14 +36,14 @@
  * INCLUDES
  ******************************************************************************/
 
-#include "PgcDemMosaicRaster.h"
+#include "GeoRaster.h"
 #include "PgcWkt.h"
 
 /******************************************************************************
  * ARCTICDEM MOSAIC RASTER CLASS
  ******************************************************************************/
 
-class ArcticDemMosaicRaster: public PgcDemMosaicRaster
+class ArcticDemMosaicRaster: public GeoRaster
 {
     public:
 
@@ -54,9 +54,6 @@ class ArcticDemMosaicRaster: public PgcDemMosaicRaster
         static RasterObject* create(lua_State* L, GeoParms* _parms)
         { return new ArcticDemMosaicRaster(L, _parms); }
 
-        int64_t getRasterDate(void) override
-        { return mosaicGetRasterDate(2023, 01, 18, 20, 23, 42); }
-
     protected:
 
         /*--------------------------------------------------------------------
@@ -64,7 +61,9 @@ class ArcticDemMosaicRaster: public PgcDemMosaicRaster
          *--------------------------------------------------------------------*/
 
         ArcticDemMosaicRaster(lua_State* L, GeoParms* _parms):
-         PgcDemMosaicRaster(L, _parms) {}
+         GeoRaster(L, _parms,
+                  std::string(_parms->asset->getPath()).append("/").append(_parms->asset->getIndex()).c_str(),
+                  getRasterDate(2023, 01, 18, 20, 23, 42)) {}
 
         void overrideTargetCRS(OGRSpatialReference& target) override
         { setCRSfromWkt(target, getArcticDemWkt2()); }
