--- plugins/pgc/plugin/PgcDemMosaicRaster.cpp
@@ -1,74 +1,0 @@
-/*
- * Copyright (c) 2021, University of Washington
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the University of Washington nor the names of its
- *    contributors may be used to endorse or promote products derived from this
- *    software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE UNIVERSITY OF WASHINGTON AND CONTRIBUTORS
- * “AS IS” AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
- * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE UNIVERSITY OF WASHINGTON OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
- * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
- * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
- * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-/******************************************************************************
- * INCLUDES
- ******************************************************************************/
-
-#include "PgcDemMosaicRaster.h"
-#include "TimeLib.h"
-
-/******************************************************************************
- * PROTECTED METHODS
- ******************************************************************************/
-
-/*----------------------------------------------------------------------------
- * Constructor
- *----------------------------------------------------------------------------*/
-PgcDemMosaicRaster::PgcDemMosaicRaster(lua_State *L, GeoParms* _parms):
-    GeoRaster(L, _parms)
-{
-    /*
-     * Pgc Mosaics use one VRT file.
-     */
-    std::string vrtFile;
-    vrtFile.append(_parms->asset->getPath()).append("/").append(_parms->asset->getIndex()).c_str();
-    openRaster(vrtFile.c_str());
-}
-
-
-/*----------------------------------------------------------------------------
- * getRasterDate
- *----------------------------------------------------------------------------*/
-int64_t PgcDemMosaicRaster::mosaicGetRasterDate(int year, int month, int day, int hour, int minute, int second)
-{
-    TimeLib::gmt_time_t gmt;
-
-    gmt.year        = year;
-    gmt.doy         = TimeLib::dayofyear(year, month, day);
-    gmt.hour        = hour;
-    gmt.minute      = minute;
-    gmt.second      = second;
-    gmt.millisecond = 0;
-
-    return TimeLib::gmt2gpstime(gmt);
-}
-
